Include ensembl info as a table & Include categorical alteration description
