+++ src/main/webapp/app/pages/annotationPage/AlterationInfo.tsx
@@ -10,10 +10,10 @@ import {
   COLOR_ICON_WITH_INFO,
   COLOR_ICON_WITHOUT_INFO,
 } from 'app/config/theme';
-import { getHighestLevelStrings } from 'app/pages/genePage/GenePage';
 import { DefaultTooltip } from 'cbioportal-frontend-commons';
 import WithSeparator from 'react-with-separator';
 import { MUTATION_EFFECT, ONCOGENICITY } from 'app/config/constants';
+import { getHighestLevelStrings } from 'app/pages/genePage/GeneInfo';
 
 export const AlterationInfo: React.FunctionComponent<{
   isPositionalAlteration: boolean;
