+++ src/main/webapp/app/pages/annotationPage/AlterationTableTabs.tsx
@@ -25,6 +25,7 @@ import { LevelOfEvidencePageLink } from 'app/shared/links/LevelOfEvidencePageLin
 import { Version } from 'app/pages/LevelOfEvidencePage';
 import {
   BiologicalVariant,
+  EnsemblGene,
   FdaAlteration,
 } from 'app/shared/api/generated/OncoKbPrivateAPI';
 import { GenePageTable } from 'app/pages/genePage/GenePageTable';
@@ -47,6 +48,7 @@ export type Column = {
 };
 export interface IEvidenceTableTabProps {
   hugoSymbol: string;
+  ensemblGenes?: EnsemblGene[];
   alteration?: IAlteration;
   cancerType?: string;
   biological: BiologicalVariant[];
@@ -304,6 +306,7 @@ export default class AlterationTableTabs extends React.Component<
             <>
               <AlterationPageLink
                 hugoSymbol={this.props.hugoSymbol}
+                ensemblGenes={this.props.ensemblGenes}
                 alteration={{
                   alteration: props.original.variant.alteration,
                   name: props.original.variant.name,
@@ -419,6 +422,7 @@ export default class AlterationTableTabs extends React.Component<
           return (
             <AlterationPageLink
               hugoSymbol={props.original.alteration.gene.hugoSymbol}
+              ensemblGenes={this.props.ensemblGenes}
               alteration={{
                 alteration: props.original.alteration.alteration,
                 name: props.original.alteration.name,
